pyproject.toml
[tool.poetry]
name = "sora"
version = "0.1.0"
description = "Simple, efficient, and powerful file and image hosting server written in Python"
authors = ["Henry \"hmont\" Monteith"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
databases = {extras = ["aiomysql"], version = "^0.9.0"}
fastapi = "^0.115.8"
redis = {extras = ["hiredis"], version = "^5.2.1"}
python-dotenv = "^1.0.1"
python-multipart = "^0.0.20"
uvicorn = "^0.34.0"
aiofiles = "^24.1.0"
bcrypt = "^4.2.1"
jinja2 = "^3.1.5"
limits = "^4.0.1"
frozendict = "^2.4.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
